Move method to get all shows to own file
